refactor: llm_chat/domain 配下のファイルを責務ごとに分割
- service/chat.py を以下のファイルに分割：
  - base.py: BaseChatService 基底クラス
  - common.py: ChatService, OpenAIChatStreamingService, get_chat_history
  - riddle.py: RiddleTask, なぞなぞ関連の機能
  - multimedia.py: DALL-E, TTS, STT 関連の機能
  - rag.py: RAG 関連の機能

- usecase/chat.py を以下のファイルに分割：
  - base.py: UseCase 基底クラス
  - common.py: LlmChatUseCase, OpenAIGptStreamingUseCase
  - riddle.py: RiddleUseCase
  - multimedia.py: OpenAIDalleUseCase, OpenAITextToSpeechUseCase, OpenAISpeechToTextUseCase
  - rag.py: OpenAIRagUseCase

- 各 __init__.py を更新し、必要なエクスポートを定義
- views.py と tests.py のインポート文を新しい構造に対応
- 既存のテストがすべてパス (12 tests OK)
