+++ llm_chat/domain/service/__init__.py
@@ -1,0 +1,33 @@
+from llm_chat.domain.service.base import BaseChatService
+from llm_chat.domain.service.common import (
+    ChatService,
+    OpenAIChatStreamingService,
+    get_chat_history,
+)
+from llm_chat.domain.service.riddle import (
+    RiddleTask,
+    RIDDLE_END_MESSAGE,
+    get_prompt,
+    create_initial_prompt,
+)
+from llm_chat.domain.service.multimedia import (
+    OpenAIDalleChatService,
+    OpenAITextToSpeechChatService,
+    OpenAISpeechToTextChatService,
+)
+from llm_chat.domain.service.rag import OpenAIRagChatService
+
+__all__ = [
+    "BaseChatService",
+    "ChatService",
+    "OpenAIChatStreamingService",
+    "get_chat_history",
+    "RiddleTask",
+    "RIDDLE_END_MESSAGE",
+    "get_prompt",
+    "create_initial_prompt",
+    "OpenAIDalleChatService",
+    "OpenAITextToSpeechChatService",
+    "OpenAISpeechToTextChatService",
+    "OpenAIRagChatService",
+]
